ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
transformers==4.35.2
sentence-transformers==2.2.2
onnxruntime==1.16.3
requests==2.31.0
aiofiles==23.2.0
pydantic==2.5.0
git+https://github.com/openai/CLIP.git
